requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0